Converting to using observable scheduled executors for testing and use them in tests
As per MANGO-1981
In preparation for MANGO-2061
